RandomTweaks

RandomTweaks.curve_fraction

RandomTweaks.curve_length

RandomTweaks.load_file_line_by_line

RandomTweaks.curve_fraction — Method

curve_fraction(f::Function,a::Real,b::Real,r::Real) -> Float64

Retunrs the position of the fraction of the length from the interval <a, b>, where the length of the curve is given by one parameter function f.

Arguments

f::Function: function of the curve,

a::Real: left-bound of the interval to be integrated on,

b::Real: right-bound of the interval to be integrated on,

r::Real: fraction of the length of the curve given by f from the left,

rtol:Real: default rtol = 1e-13, relative tolerance,

Keywords

Returns

Float64: position of the fraction of the length from the interval <a, b>.

Throws

RandomTweaks.curve_length — Method

curve_length(f::Function,a::Real,b::Real;rtol=1e-13) -> Float64

Calculates the length of the curve given by f, where f is function of one parameter, on interval <a, b> of said parameter.

Arguments

f::Function: function of the curve to be measured,

a::Real: left-bound of the interval to be integrated on,

b::Real: right-bound of the interval to be integrated on,

rtol::Real: default rtol = 1e-13, relative tolerance,

Keywords

Returns

Float64: length of the curve defined by the f on the interval <a, b>.

Throws

RandomTweaks.load_file_line_by_line — Method

load_file_line_by_line(filename::AbstractString) -> Array{SubString{String},1}

Load and return the content of the text file filename in Array{SubString{String},1}, i.e., one element for each line.

Arguments

filename::AbstractString: the filename of the file to load,

Keywords

Returns

Array{SubString{String},1}: the array of the file content,

Throws

Error: in the case of file does not exist.